+++ app/tier2.go
@@ -30,8 +30,7 @@ type Tier2Config struct {
 	WASMExtensions  []wasm.WASMExtensioner
 	PipelineOptions []pipeline.PipelineOptioner
 
-	RequestStats bool
-	Tracing      bool
+	Tracing bool
 }
 
 type Tier2App struct {
@@ -80,10 +79,6 @@ func (a *Tier2App) Run() error {
 
 	if a.config.Tracing {
 		opts = append(opts, service.WithModuleExecutionTracing())
-	}
-
-	if a.config.RequestStats {
-		opts = append(opts, service.WithRequestStats())
 	}
 
 	svc := service.NewTier2(