+++ pipeline/exec/module_executor_test.go
@@ -6,9 +6,12 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"go.uber.org/zap"
 
+	"github.com/streamingfast/substreams/metrics"
 	pbssinternal "github.com/streamingfast/substreams/pb/sf/substreams/intern/v2"
 	pbsubstreams "github.com/streamingfast/substreams/pb/sf/substreams/v1"
+	"github.com/streamingfast/substreams/reqctx"
 	"github.com/streamingfast/substreams/storage/execout"
 )
 
@@ -91,6 +94,8 @@ func (t *MockModuleExecutor) lastExecutionStack() []string {
 
 func TestModuleExecutorRunner_Run_HappyPath(t *testing.T) {
 	ctx := context.Background()
+
+	ctx = reqctx.WithReqStats(ctx, metrics.NewReqStats(&metrics.Config{}, zap.NewNop()))
 	executor := &MockModuleExecutor{
 		name: "test",
 		RunFunc: func(ctx context.Context, reader execout.ExecutionOutputGetter) (out []byte, moduleOutputData *pbssinternal.ModuleOutput, err error) {