+++ wasm/call.go
@@ -3,10 +3,12 @@ package wasm
 import (
 	"fmt"
 	"math/big"
+	"time"
 
 	"github.com/dustin/go-humanize"
 	"github.com/shopspring/decimal"
 
+	"github.com/streamingfast/substreams/metrics"
 	pbsubstreams "github.com/streamingfast/substreams/pb/sf/substreams/v1"
 	"github.com/streamingfast/substreams/storage/store"
 )
@@ -28,13 +30,15 @@ type Call struct {
 	Logs           []string
 	LogsByteCount  uint64
 	ExecutionStack []string
+	stats          *metrics.Stats
 }
 
-func NewCall(clock *pbsubstreams.Clock, moduleName string, entrypoint string, arguments []Argument) *Call {
+func NewCall(clock *pbsubstreams.Clock, moduleName string, entrypoint string, stats *metrics.Stats, arguments []Argument) *Call {
 	call := &Call{
 		Clock:      clock,
 		ModuleName: moduleName,
 		Entrypoint: entrypoint,
+		stats:      stats,
 	}
 
 	for _, input := range arguments {
@@ -113,30 +117,36 @@ func (c *Call) ReachedLogsMaxByteCount() bool {
 }
 
 func (c *Call) DoSet(ord uint64, key string, value []byte) {
+	defer c.stats.RecordModuleWasmStoreWrite(c.ModuleName, time.Since(time.Now()))
 	c.validateSimple("set", pbsubstreams.Module_KindStore_UPDATE_POLICY_SET, key)
 	c.outputStore.SetBytes(ord, key, value)
 }
 func (c *Call) DoSetIfNotExists(ord uint64, key string, value []byte) {
+	defer c.stats.RecordModuleWasmStoreWrite(c.ModuleName, time.Since(time.Now()))
 	c.validateSimple("set_if_not_exists", pbsubstreams.Module_KindStore_UPDATE_POLICY_SET_IF_NOT_EXISTS, key)
 	c.outputStore.SetBytesIfNotExists(ord, key, value)
 }
 func (c *Call) DoAppend(ord uint64, key string, value []byte) {
+	defer c.stats.RecordModuleWasmStoreWrite(c.ModuleName, time.Since(time.Now()))
 	c.validateSimple("append", pbsubstreams.Module_KindStore_UPDATE_POLICY_APPEND, key)
 	if err := c.outputStore.Append(ord, key, value); err != nil {
 		c.ReturnError(fmt.Errorf("appending to store: %w", err))
 	}
 }
 func (c *Call) DoDeletePrefix(ord uint64, prefix string) {
+	defer c.stats.RecordModuleWasmStoreDeletePrefix(c.ModuleName, time.Since(time.Now()))
 	c.traceStateWrites("delete_prefix", prefix)
 	c.outputStore.DeletePrefix(ord, prefix)
 }
 func (c *Call) DoAddBigInt(ord uint64, key string, value string) {
+	defer c.stats.RecordModuleWasmStoreWrite(c.ModuleName, time.Since(time.Now()))
 	c.validateWithValueType("add_bigint", pbsubstreams.Module_KindStore_UPDATE_POLICY_ADD, "bigint", key)
 
 	toAdd, _ := new(big.Int).SetString(value, 10)
 	c.outputStore.SumBigInt(ord, key, toAdd)
 }
 func (c *Call) DoAddBigDecimal(ord uint64, key string, value string) {
+	defer c.stats.RecordModuleWasmStoreWrite(c.ModuleName, time.Since(time.Now()))
 	c.validateWithTwoValueTypes("add_bigdecimal", pbsubstreams.Module_KindStore_UPDATE_POLICY_ADD, "bigdecimal", "bigfloat", key)
 
 	toAdd, err := decimal.NewFromString(string(value))
@@ -146,27 +156,33 @@ func (c *Call) DoAddBigDecimal(ord uint64, key string, value string) {
 	c.outputStore.SumBigDecimal(ord, key, toAdd.Truncate(34))
 }
 func (c *Call) DoAddInt64(ord uint64, key string, value int64) {
+	defer c.stats.RecordModuleWasmStoreWrite(c.ModuleName, time.Since(time.Now()))
 	c.validateWithValueType("add_int64", pbsubstreams.Module_KindStore_UPDATE_POLICY_ADD, "int64", key)
 	c.outputStore.SumInt64(ord, key, value)
 }
 func (c *Call) DoAddFloat64(ord uint64, key string, value float64) {
+	defer c.stats.RecordModuleWasmStoreWrite(c.ModuleName, time.Since(time.Now()))
 	c.validateWithValueType("add_float64", pbsubstreams.Module_KindStore_UPDATE_POLICY_ADD, "float64", key)
 	c.outputStore.SumFloat64(ord, key, value)
 }
 func (c *Call) DoSetMinInt64(ord uint64, key string, value int64) {
+	defer c.stats.RecordModuleWasmStoreWrite(c.ModuleName, time.Since(time.Now()))
 	c.validateWithValueType("set_min_int64", pbsubstreams.Module_KindStore_UPDATE_POLICY_MIN, "int64", key)
 	c.outputStore.SetMinInt64(ord, key, value)
 }
 func (c *Call) DoSetMinBigInt(ord uint64, key string, value string) {
+	defer c.stats.RecordModuleWasmStoreWrite(c.ModuleName, time.Since(time.Now()))
 	c.validateWithValueType("set_min_bigint", pbsubstreams.Module_KindStore_UPDATE_POLICY_MIN, "bigint", key)
 	toSet, _ := new(big.Int).SetString(value, 10)
 	c.outputStore.SetMinBigInt(ord, key, toSet)
 }
 func (c *Call) DoSetMinFloat64(ord uint64, key string, value float64) {
+	defer c.stats.RecordModuleWasmStoreWrite(c.ModuleName, time.Since(time.Now()))
 	c.validateWithValueType("set_min_float64", pbsubstreams.Module_KindStore_UPDATE_POLICY_MIN, "float64", key)
 	c.outputStore.SetMinFloat64(ord, key, value)
 }
 func (c *Call) DoSetMinBigDecimal(ord uint64, key string, value string) {
+	defer c.stats.RecordModuleWasmStoreWrite(c.ModuleName, time.Since(time.Now()))
 	c.validateWithTwoValueTypes("set_min_bigdecimal", pbsubstreams.Module_KindStore_UPDATE_POLICY_MIN, "bigdecimal", "bigfloat", key)
 	toAdd, err := decimal.NewFromString(value)
 	if err != nil {
@@ -175,20 +191,24 @@ func (c *Call) DoSetMinBigDecimal(ord uint64, key string, value string) {
 	c.outputStore.SetMinBigDecimal(ord, key, toAdd.Truncate(34))
 }
 func (c *Call) DoSetMaxInt64(ord uint64, key string, value int64) {
+	defer c.stats.RecordModuleWasmStoreWrite(c.ModuleName, time.Since(time.Now()))
 	c.validateWithValueType("set_max_int64", pbsubstreams.Module_KindStore_UPDATE_POLICY_MAX, "int64", key)
 	c.outputStore.SetMaxInt64(ord, key, value)
 }
 func (c *Call) DoSetMaxBigInt(ord uint64, key string, value string) {
+	defer c.stats.RecordModuleWasmStoreWrite(c.ModuleName, time.Since(time.Now()))
 	c.validateWithValueType("set_max_bigint", pbsubstreams.Module_KindStore_UPDATE_POLICY_MAX, "bigint", key)
 	toSet, _ := new(big.Int).SetString(value, 10)
 	c.outputStore.SetMaxBigInt(ord, key, toSet)
 
 }
 func (c *Call) DoSetMaxFloat64(ord uint64, key string, value float64) {
+	defer c.stats.RecordModuleWasmStoreWrite(c.ModuleName, time.Since(time.Now()))
 	c.validateWithValueType("set_max_float64", pbsubstreams.Module_KindStore_UPDATE_POLICY_MAX, "float64", key)
 	c.outputStore.SetMaxFloat64(ord, key, value)
 }
 func (c *Call) DoSetMaxBigDecimal(ord uint64, key string, value string) {
+	defer c.stats.RecordModuleWasmStoreWrite(c.ModuleName, time.Since(time.Now()))
 	c.validateWithTwoValueTypes("set_max_bigdecimal", pbsubstreams.Module_KindStore_UPDATE_POLICY_MAX, "bigdecimal", "bigfloat", key)
 	toAdd, err := decimal.NewFromString(value)
 	if err != nil {
@@ -198,6 +218,7 @@ func (c *Call) DoSetMaxBigDecimal(ord uint64, key string, value string) {
 }
 
 func (c *Call) DoGetAt(storeIndex int, ord uint64, key string) (value []byte, found bool) {
+	defer c.stats.RecordModuleWasmStoreRead(c.ModuleName, time.Since(time.Now()))
 	c.validateStoreIndex(storeIndex, "get_at")
 	readStore := c.inputStores[storeIndex]
 	c.traceStateReads("get_at", storeIndex, found, key)
@@ -205,6 +226,7 @@ func (c *Call) DoGetAt(storeIndex int, ord uint64, key string) (value []byte, fo
 }
 
 func (c *Call) DoHasAt(storeIndex int, ord uint64, key string) (found bool) {
+	defer c.stats.RecordModuleWasmStoreRead(c.ModuleName, time.Since(time.Now()))
 	c.validateStoreIndex(storeIndex, "has_at")
 	readStore := c.inputStores[storeIndex]
 	c.traceStateReads("has_at", storeIndex, found, key)
@@ -212,6 +234,7 @@ func (c *Call) DoHasAt(storeIndex int, ord uint64, key string) (found bool) {
 }
 
 func (c *Call) DoGetFirst(storeIndex int, key string) (value []byte, found bool) {
+	defer c.stats.RecordModuleWasmStoreRead(c.ModuleName, time.Since(time.Now()))
 	c.validateStoreIndex(storeIndex, "get_first")
 	readStore := c.inputStores[storeIndex]
 	c.traceStateReads("get_first", storeIndex, found, key)
@@ -219,6 +242,7 @@ func (c *Call) DoGetFirst(storeIndex int, key string) (value []byte, found bool)
 }
 
 func (c *Call) DoHasFirst(storeIndex int, key string) (found bool) {
+	defer c.stats.RecordModuleWasmStoreRead(c.ModuleName, time.Since(time.Now()))
 	c.validateStoreIndex(storeIndex, "has_first")
 	readStore := c.inputStores[storeIndex]
 	c.traceStateReads("has_first", storeIndex, found, key)
@@ -226,6 +250,7 @@ func (c *Call) DoHasFirst(storeIndex int, key string) (found bool) {
 }
 
 func (c *Call) DoGetLast(storeIndex int, key string) (value []byte, found bool) {
+	defer c.stats.RecordModuleWasmStoreRead(c.ModuleName, time.Since(time.Now()))
 	c.validateStoreIndex(storeIndex, "get_last")
 	readStore := c.inputStores[storeIndex]
 	c.traceStateReads("get_last", storeIndex, found, key)
@@ -233,6 +258,7 @@ func (c *Call) DoGetLast(storeIndex int, key string) (value []byte, found bool) 
 }
 
 func (c *Call) DoHasLast(storeIndex int, key string) (found bool) {
+	defer c.stats.RecordModuleWasmStoreRead(c.ModuleName, time.Since(time.Now()))
 	c.validateStoreIndex(storeIndex, "has_last")
 	readStore := c.inputStores[storeIndex]
 	c.traceStateReads("has_last", storeIndex, found, key)